Add argument to be able to exclude directories

+++ index.js
@@ -7,24 +7,45 @@ const yargs = require("yargs");
 
 const cwd = process.cwd();
 const argv = yargs.argv;
+const excludes = argv.exclude ? argv.exclude.split(",") : [];
 
-argv._.forEach(moduleName => {
+argv._.forEach((moduleName) => {
   console.log("Copying [" + moduleName + "]");
   const modulePath = globalRequire.resolve(moduleName);
 
   const fullDestPath = path.join(cwd, "node_modules", moduleName);
 
-  const destPath = fs.lstatSync(fullDestPath);
-  if (destPath.isSymbolicLink()) {
-    console.log("Was already simlinked [" + moduleName + "]. Removing the simlink first");
-    fs.unlinkSync(fullDestPath)
+  if (fs.existsSync(fullDestPath)) {
+    const destPath = fs.lstatSync(fullDestPath);
+    if (destPath.isSymbolicLink()) {
+      console.log(
+        "Was already simlinked [" + moduleName + "]. Removing the simlink first"
+      );
+      fs.unlinkSync(fullDestPath);
+    }
   }
 
-  fs.copy(modulePath, fullDestPath, { dereference: true }, error => {
-    if (error) {
-      console.log("Cannot copy [" + moduleName + "]. Error: " + error.message);
-    } else {
-      console.log("Copy of [" + moduleName + "] finished.");
+  fs.copy(
+    modulePath,
+    fullDestPath,
+    {
+      dereference: true,
+      filter: (path) => {
+        if (excludes.length > 0 && fs.lstatSync(path).isDirectory()) {
+          const  lastPart = path.split("/").pop();
+          return (excludes.indexOf(lastPart) === -1);
+        }
+        return true;
+      },
+    },
+    (error) => {
+      if (error) {
+        console.log(
+          "Cannot copy [" + moduleName + "]. Error: " + error.message
+        );
+      } else {
+        console.log("Copy of [" + moduleName + "] finished.");
+      }
     }
-  });
+  );
 });
